+++ cli/pg_datanymizer/src/options.rs
@@ -11,9 +11,10 @@ pub struct Options {
     #[structopt(
         short = "c",
         long = "config",
-        help = "Path to config file. Default: ./config.yml"
+        help = "Path to config file",
+        default_value = "./config.yml"
     )]
-    pub config: Option<String>,
+    pub config: String,
     #[structopt(
         short = "f",
         long = "file",
@@ -32,23 +33,23 @@ pub struct Options {
     #[structopt(
         short = "h",
         long = "host",
-        help = "database server host or socket directory",
+        help = "Database server host or socket directory",
         default_value = "localhost"
     )]
     pub host: String,
-    #[structopt(short = "p", long = "port", help = "database server port number")]
+    #[structopt(
+        short = "p",
+        long = "port",
+        help = "Database server port number [default: 5432]"
+    )]
     pub port: Option<u16>,
     #[structopt(
         short = "U",
         long = "username",
-        help = "connect as specified database user"
+        help = "Connect as specified database user"
     )]
     pub username: Option<String>,
-    #[structopt(
-        short = "W",
-        long = "password",
-        help = "force password prompt (should happen automatically)"
-    )]
+    #[structopt(short = "W", long = "password", help = "User password")]
     pub password: Option<String>,
     #[structopt(
         long = "pg_dump",
@@ -56,13 +57,24 @@ pub struct Options {
         default_value = "pg_dump"
     )]
     pub pg_dump_location: String,
+
+    #[structopt(
+        long = "accept_invalid_hostnames",
+        help = "Accept or not invalid hostnames when using SSL [default: false]"
+    )]
+    pub accept_invalid_hostnames: Option<bool>,
+    #[structopt(
+        long = "accept_invalid_certs",
+        help = "Accept or not invalid certificates (e.g., self-signed) when using SSL [default: false]"
+    )]
+    pub accept_invalid_certs: Option<bool>,
 }
 
 impl Options {
-    pub fn database_url(&self) -> Result<String> {
+    pub fn database_url(&self) -> Result<Url> {
         if let Ok(url) = Url::parse(self.database.as_str()) {
             if url.scheme() == "postgres" {
-                return Ok(url.to_string());
+                return Ok(url);
             } else {
                 return Err(anyhow!("Scheme url error"));
             }
@@ -70,7 +82,7 @@ impl Options {
         self.build_url(Some(self.database.to_string()).filter(|x| !x.is_empty()))
     }
 
-    fn build_url(&self, override_db_name: Option<String>) -> Result<String> {
+    fn build_url(&self, override_db_name: Option<String>) -> Result<Url> {
         let db_name = override_db_name.unwrap_or_else(|| self.db_name.clone());
         if db_name.is_empty() {
             return Err(anyhow!("No one database passed"));
@@ -88,19 +100,19 @@ impl Options {
 
         url.set_path(&db_name);
 
-        Ok(url.to_string())
+        Ok(url)
     }
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::Options;
 
     #[test]
     fn parse_empty_config() {
         let cfg = Options {
             database: "postgres://hostname/test".to_string(),
-            config: None,
+            config: "./config.yml".to_string(),
             file: None,
             db_name: "test".to_string(),
             host: "localhost".to_string(),
@@ -108,17 +120,19 @@ mod test {
             username: None,
             password: None,
             pg_dump_location: "pg_dump".to_string(),
+            accept_invalid_hostnames: Some(false),
+            accept_invalid_certs: Some(false),
         };
 
         let expected = "postgres://hostname/test".to_string();
-        assert_eq!(cfg.database_url().unwrap(), expected);
+        assert_eq!(cfg.database_url().unwrap().to_string(), expected);
     }
 
     #[test]
     fn parse_empty_url() {
         let cfg1 = Options {
             database: String::default(),
-            config: None,
+            config: "./config.yml".to_string(),
             file: None,
             db_name: "test".to_string(),
             host: "hostname".to_string(),
@@ -126,6 +140,8 @@ mod test {
             username: None,
             password: None,
             pg_dump_location: "pg_dump".to_string(),
+            accept_invalid_hostnames: Some(false),
+            accept_invalid_certs: Some(false),
         };
 
         let cfg2 = Options {
@@ -145,9 +161,9 @@ mod test {
         let expected2 = "postgres://hostname:5433/test".to_string();
         let expected3 = "postgres://test_user@hostname:5433/test".to_string();
         let expected4 = "postgres://test_user:pass@hostname:5433/test".to_string();
-        assert_eq!(cfg1.database_url().unwrap(), expected1);
-        assert_eq!(cfg2.database_url().unwrap(), expected2);
-        assert_eq!(cfg3.database_url().unwrap(), expected3);
-        assert_eq!(cfg4.database_url().unwrap(), expected4);
+        assert_eq!(cfg1.database_url().unwrap().to_string(), expected1);
+        assert_eq!(cfg2.database_url().unwrap().to_string(), expected2);
+        assert_eq!(cfg3.database_url().unwrap().to_string(), expected3);
+        assert_eq!(cfg4.database_url().unwrap().to_string(), expected4);
     }
 }