+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### 🚀 Added
+- Options to ignore invalid certificates and hostnames when using SSL
+  [#64](https://github.com/datanymizer/datanymizer/pull/64) ([@evgeniy-r](https://github.com/evgeniy-r))
 - Improve docs [#67](https://github.com/datanymizer/datanymizer/pull/67) ([@evgeniy-r](https://github.com/evgeniy-r))
 - Add the basic SSL support for Postgres [#61](https://github.com/datanymizer/datanymizer/pull/61)
   ([@evgeniy-r](https://github.com/evgeniy-r))